SimpleBasicTracker/requirements.txt
flask==2.3.3
requests==2.31.0
pandas==2.0.3
openpyxl==3.1.2
pytz==2023.3
gspread==5.11.3
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
beautifulsoup4>=4.12.0
schedule==1.2.0
Flask>=2.3.0
Werkzeug>=2.3.0
psutil>=5.9.0
python-dotenv==1.0.0 